+++ gen/armor.py
@@ -184,9 +184,9 @@ class GroundConflictGenerator:
 
         for dcs_group, group in ally_groups:
 
-            if hasattr(group.unit_type, 'eplrs'):
-                if group.unit_type.eplrs:
-                    group.points[0].tasks.append(EPLRS(group.id))
+            if hasattr(group.units[0], 'eplrs'):
+                if group.units[0].eplrs:
+                    dcs_group.points[0].tasks.append(EPLRS(dcs_group.id))
 
             if group.role == CombatGroupRole.ARTILLERY:
                 # Fire on any ennemy in range